+++ src/events.js
@@ -3,5 +3,7 @@ export const EVENT_READTHEDOCS_SEARCH_HIDE = "readthedocs-search-hide";
 export const EVENT_READTHEDOCS_DOCDIFF_ADDED_REMOVED_SHOW =
   "readthedocs-docdiff-added-removed-show";
 export const EVENT_READTHEDOCS_DOCDIFF_HIDE = "readthedocs-docdiff-hide";
+export const EVENT_READTHEDOCS_FLYOUT_SHOW = "readthedocs-flyout-show";
+export const EVENT_READTHEDOCS_FLYOUT_HIDE = "readthedocs-flyout-hide";
 export const EVENT_READTHEDOCS_ADDONS_DATA_READY =
   "readthedocs-addons-data-ready";