+++ web-test-runner.config.mjs
@@ -1,10 +1,12 @@
 import { importMapsPlugin } from "@web/dev-server-import-maps";
 import { fromRollup, rollupAdapter } from "@web/dev-server-rollup";
+import rollupReplace from "@rollup/plugin-replace";
 import rollupCommonjs from "@rollup/plugin-commonjs";
 import rollupImage from "@rollup/plugin-image";
 import rollupJson from "@rollup/plugin-json";
 import rollupLitCss from "rollup-plugin-lit-css";
 
+const replace = fromRollup(rollupReplace);
 const pluginCommonjs = fromRollup(rollupCommonjs);
 
 export default {
@@ -26,6 +28,13 @@ export default {
       esmExternals: true,
       defaultIsModuleExports: true,
       requireReturnsDefault: true,
+    }),
+    // https://modern-web.dev/docs/dev-server/writing-plugins/examples/#environment-variables
+    replace({
+      preventAssignment: true,
+      include: ["src/**/*.js"],
+      WEBPACK_IS_TESTING: JSON.stringify(true),
+      WEBPACK_IS_PRODUCTION: JSON.stringify(false),
     }),
   ],
 };